2007-06-06 Hans Boehm <Hans.Boehm@hp.com> 	* src/atomic_ops/sysdeps/msftc/x86_64.h: New file. 	* src/atomic_ops.h: Add test for msftc/x86_64.h. 	* src/atomic_ops/sysdeps/msftc/x86.h: Complain for _WIN64. 	* src/atomic_ops/sysdeps/Makefile.am: Add x86_64.h. 	* src/atomic_ops/sysdeps/Makefile.in: Regenerate. 	* src/atomic_ops/sysdeps/aligned_atomic_load_store.h, 	  src/atomic_ops/sysdeps/int_aligned_atomic_load_store.h, 	  src/atomic_ops/sysdeps/short_aligned_atomic_load_store.h: 	  Replace unsigned long cast with size_t.

+++ src/atomic_ops/sysdeps/aligned_atomic_load_store.h
@@ -28,7 +28,7 @@
 AO_INLINE AO_t
 AO_load(volatile AO_t *addr)
 {
-  assert(((unsigned long)addr & (sizeof(AO_t) - 1)) == 0);
+  assert(((size_t)addr & (sizeof(AO_t) - 1)) == 0);
   /* Cast away the volatile for architectures where		*/
   /* volatile adds barrier semantics.				*/
   return *(AO_t *)addr;
@@ -39,7 +39,7 @@ AO_load(volatile AO_t *addr)
 AO_INLINE void
 AO_store(volatile AO_t *addr, AO_t new_val)
 {
-  assert(((unsigned long)addr & (sizeof(AO_t) - 1)) == 0);
+  assert(((size_t)addr & (sizeof(AO_t) - 1)) == 0);
   (*(AO_t *)addr) = new_val;
 }
