+++ src/atomic_ops/sysdeps/short_aligned_atomic_load_store.h
@@ -28,7 +28,7 @@
 AO_INLINE unsigned short
 AO_short_load(volatile unsigned short *addr)
 {
-  assert(((unsigned long)addr & (sizeof(unsigned short) - 1)) == 0);
+  assert(((size_t)addr & (sizeof(unsigned short) - 1)) == 0);
   /* Cast away the volatile for architectures like IA64 where	*/
   /* volatile adds barrier semantics.				*/
   return (*(unsigned short *)addr);
@@ -39,7 +39,7 @@ AO_short_load(volatile unsigned short *addr)
 AO_INLINE void
 AO_short_store(volatile unsigned short *addr, unsigned short new_val)
 {
-  assert(((unsigned long)addr & (sizeof(unsigned short) - 1)) == 0);
+  assert(((size_t)addr & (sizeof(unsigned short) - 1)) == 0);
   (*(unsigned short *)addr) = new_val;
 }
